requirements.txt
Flask
Flask-Bcrypt
Flask-SQLAlchemy
Flask-Login
Flask-Migrate
psycopg2-binary
openpyxl
gunicorn
python-dotenv
Werkzeug
pytest